Update ColorSensorTest.java
Made the code run 3 different color tests at the start of the match. Once I figure out the motors and driving, I should be able to program the robot to do different things based on where it senses the color. WIP - lmk if u get errors and send me a pic of them.